♻️ Merge auth-cookies into helpers/server/cookies.ts
Single source of truth for getCookieOptions (with domain) and
buildSessionCookies. Fixes logout where cookieStore.delete() used
a different getCookieOptions (without domain) than the one used
to set the cookies, causing silent logout failure.
